update dependencies, move to esm
