+++ package.json
@@ -1,32 +1,33 @@
 {
-  "name": "@mikemcbride/dad-jokes",
-  "version": "1.0.0",
-  "description": "Everybody loves dad jokes... right?",
-  "bin": "cli.js",
-  "engines": "node >= 4",
-  "main": "index.js",
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/mikemcbride/dad-jokes.git"
-  },
-  "files": [
-    "index.js",
-    "cli.js",
-    "jokes.json"
-  ],
-  "keywords": [
-    "cli",
-    "cli-app",
-    "dad-joke",
-    "joke",
-    "jokes",
-    "dad",
-    "funny"
-  ],
-  "author": "mikemcbride",
-  "license": "MIT",
-  "dependencies": {
-    "meow": "^4.0.0",
-    "unique-random-array": "^1.0.0"
-  }
+    "name": "@mikemcbride/dad-jokes",
+    "version": "2.0.0",
+    "description": "Everybody loves dad jokes... right?",
+    "bin": "cli.js",
+    "engines": "node >= 12",
+    "type": "module",
+    "main": "index.js",
+    "repository": {
+        "type": "git",
+        "url": "git+https://github.com/mikemcbride/dad-jokes.git"
+    },
+    "files": [
+        "index.js",
+        "cli.js",
+        "jokes.js"
+    ],
+    "keywords": [
+        "cli",
+        "cli-app",
+        "dad-joke",
+        "joke",
+        "jokes",
+        "dad",
+        "funny"
+    ],
+    "author": "mikemcbride",
+    "license": "MIT",
+    "dependencies": {
+        "meow": "^10.1.3",
+        "unique-random-array": "^1.0.0"
+    }
 }
